+++ tests/unit_tests/providers/semantic_scholar_test.py
@@ -1,9 +1,11 @@
 from datetime import date
-from sciety_labs.providers.semantic_scholar.semantic_scholar import (
-    _get_recommendation_request_payload_for_paper_ids_or_external_ids,
+from sciety_labs.providers.semantic_scholar.utils import (
     _iter_article_recommendation_from_recommendation_response_json
 )
-from sciety_labs.providers.semantic_scholar.utils import SEMANTIC_SCHOLAR_PAPER_ID_EXT_REF_ID
+from sciety_labs.providers.semantic_scholar.utils import (
+    SEMANTIC_SCHOLAR_PAPER_ID_EXT_REF_ID,
+    _get_recommendation_request_payload_for_paper_ids_or_external_ids
+)
 
 
 DOI_1 = '10.1101/doi1'
